jit: measure the spill plan on both targets
Moves the plan measurement out of `asm_dump`, which is aarch64-only, into
`spill`'s own tests, which both targets compile. It never needed an encoder —
only lowering, isel and a MachineEnv.

That makes the interesting comparison possible, and it changes the reading.
Braun09 measured on x86 with 7 registers, chosen deliberately to stress
spilling; aarch64 gives us 20, where mix2's pressure barely exceeds the file and
there is little for a better policy to win. x86-64's 13 is much closer to the
paper's setting:

  x86-64 (13 regs)   mix   38 ops (w 344)  vs 69 (w 483)    -45% ops, -29% weighted
                     mix2 279 ops (w 2169) vs 330 (w 2715)  -15% ops, -20% weighted
  aarch64 (20 regs)  mix2 193 ops (w 1759) vs 210 (w 1722)   -8% ops,  +2% weighted

So the pass wins where there is real pressure and is a wash where there is not,
which is the shape the paper's own choice of 7 registers implies.

The "a function that fits needs no spill code" assertion now keys on the
measured peak rather than on which benchmark it is, since that differs by
target: mix peaks at 19, which fits aarch64's 20 and not x86-64's 13. Keyed on
the filename it asserted something false on x64.
