+++ src/jit/backend/spill.rs
@@ -868,4 +868,148 @@ mod tests {
             "a value with no further use must be dropped, not stored"
         );
     }
+
+    /// The spiller, on real functions.
+    ///
+    /// Three properties, each of which caught a real bug while this was being written:
+    ///
+    /// - **Pressure is within the register file everywhere.** Every claim downstream —
+    ///   above all that assignment can then proceed without spilling at all — rests on
+    ///   this.
+    /// - **Nothing is reloaded that was never stored.** A reload reads a slot; if no
+    ///   store ever wrote that slot it reads garbage. This is the invariant that
+    ///   exposed dead values occupying `W`, jump arguments being dropped at block
+    ///   exits, and block parameters being reloaded rather than delivered by the edge.
+    /// - **A function that fits needs no spill code.** `is_prime` and `mix` both peak
+    ///   below the register file, so a correct spiller must leave them completely
+    ///   alone — and does, which is a claim about the whole pipeline agreeing with an
+    ///   allocator that reached the same answer by an entirely different route.
+    #[test]
+    fn the_spill_plan_is_within_k_and_coherent() {
+        use crate::Lua;
+        use crate::jit::backend::isel::select;
+        use crate::jit::backend::regalloc::allocate;
+        use crate::jit::backend::target::{annotate, machine_env};
+        use crate::jit::backend::{nextuse, order, spillcost};
+        use crate::jit::frontend::lower::lower;
+        use crate::jit::ir::ty::{Rep, Ty, TypeSet};
+
+        const INT: Ty = Ty::new(Rep::Val, TypeSet::INT);
+
+        eprintln!("\n=== spill plan ===");
+        for (file, chunk) in [("primes", "primes"), ("mix", "mix"), ("mix2", "mix2")] {
+            let source = std::fs::read_to_string(format!("test-files/{file}.lua")).unwrap();
+            let mut lua = Lua::new();
+            lua.load_all();
+            lua.enter(|ctx| {
+                let c = ctx.load(&source, Some(chunk)).expect("compile");
+                let proto = c.as_lua().expect("lua closure").proto.prototypes[0];
+                let func = lower(proto, 0, vec![INT]).expect("lower");
+                let mut m = select(&func).expect("isel");
+                annotate(&mut m);
+
+                let env = machine_env();
+                let layout = order::compute(&m).expect("reducible");
+                let nu = nextuse::analyze(&m, &layout);
+                let plan = plan(&m, &layout, &nu, &env);
+
+                for class in RegClass::ALL {
+                    let k = env.order(class).len();
+                    for b in 0..m.num_blocks() {
+                        for (what, set) in [("entry", &plan.w_entry[b]), ("exit", &plan.w_exit[b])]
+                        {
+                            let n = set.iter().filter(|&&v| m.class(v) == class).count();
+                            assert!(
+                                n <= k,
+                                "{file} mb{b} {what}: {n} {class:?} values in registers, k = {k}"
+                            );
+                        }
+                    }
+                }
+
+                let stored: Vec<VReg> = plan
+                    .spill_before
+                    .iter()
+                    .flatten()
+                    .chain(plan.edge_spill.values().flatten())
+                    .copied()
+                    .collect();
+                for v in plan
+                    .reload_before
+                    .iter()
+                    .flatten()
+                    .chain(plan.edge_reload.values().flatten())
+                {
+                    assert!(
+                        stored.contains(v) || m.remat(*v).is_some(),
+                        "{file}: v{} is reloaded but never stored — the slot it reads \
+                         was never written",
+                        v.0
+                    );
+                }
+
+                // The plan's own cost, weighted the same way `spillcost` weights the
+                // allocator's output, so the two are comparable. Two conventions have to
+                // match for that: an edge's code lands in the predecessor, so it is
+                // charged at the predecessor's depth; and a rematerialized value is
+                // replayed rather than loaded, so like `spillcost` it is counted apart
+                // from memory traffic rather than as some of it.
+                let mem = |v: &&VReg| m.remat(**v).is_none();
+                let ops = plan
+                    .reload_before
+                    .iter()
+                    .chain(&plan.spill_before)
+                    .chain(plan.edge_reload.values())
+                    .chain(plan.edge_spill.values())
+                    .flatten()
+                    .filter(mem)
+                    .count();
+                let mut weighted = 0u64;
+                let mut depth_of_inst = vec![0u32; m.num_insts()];
+                for &b in &layout.order {
+                    for &i in m.block_insts(b) {
+                        depth_of_inst[i] = layout.depth[b.0 as usize];
+                    }
+                }
+                for (i, (r, sp)) in plan
+                    .reload_before
+                    .iter()
+                    .zip(&plan.spill_before)
+                    .enumerate()
+                {
+                    let n = r.iter().chain(sp).filter(mem).count();
+                    weighted += n as u64 * 10u64.pow(depth_of_inst[i]);
+                }
+                for (&(p, _), vs) in plan.edge_reload.iter().chain(plan.edge_spill.iter()) {
+                    let n = vs.iter().filter(mem).count();
+                    weighted += n as u64 * 10u64.pow(layout.depth[p.0 as usize]);
+                }
+
+                let now = spillcost::measure(&m, &allocate(&m, &env).expect("allocate"))
+                    .expect("reducible");
+                eprintln!(
+                    "  {file:6} plan {} ops (weighted {weighted})   vs today {} ops \
+                     (weighted {})",
+                    ops,
+                    now.total(),
+                    now.weighted,
+                );
+
+                // A function whose peak pressure fits must come out untouched. Keyed
+                // on the measured peak rather than on which benchmark it is, because
+                // that differs by target: mix peaks at 19, which fits aarch64's 20
+                // registers and does not fit x86-64's 13.
+                let fits = RegClass::ALL.iter().all(|&c| {
+                    let k = env.order(c).len() as u32;
+                    nu.pressure.iter().all(|p| p[c as usize] <= k)
+                });
+                if fits {
+                    assert_eq!(
+                        ops, 0,
+                        "{file} peaks below the register file, so it needs no spill code"
+                    );
+                }
+            });
+        }
+    }
 }
